requirements.txt
SpeechRecognition
pyttsx3
pyaudio
wikipedia
requests
google-genai
